small update

+++ docs/cli_inputoutput.py
@@ -1,6 +1,6 @@
 # cli_inputoutput.py
-a = int(input("A=")) # ①
-b = int(input("B=")) # ②
+a = float(input("A="))  # ①
+b = float(input("B="))  # ②
 if b != 0:
   print("A/B = ", a / b)  # ③
 else:
